Small change to api url addresses
I changed the description file so that the port that the backend runs is visible when it describes the api requests. As a result frontend team can easily see which port to send post requests to without understanding the backend code. Thus frontend team can fully concentrate on implementing frontend features.