Maps, Signup and Communities Page Improvements
In that commit I corrected Maps functionality so that a user can now
open the map and choose, search, drag locations.
Now the SignUp pags contain all neccesary inputs like city, name,
surname, birthdate and pplink.
The Communities Page now is connected to backend so that the communities
on the server are listed in that page.

[Ticket: Muhammet Çavuş]